trunk:svn_tom:fix bug 0016016: edit maxInclusive and save,pop up error message
git-svn-id: http://talendforge.org/svn/tom/trunk@49165 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.mdm.workbench/src/com/amalto/workbench/actions/XSDEditFacetAction.java
@@ -414,7 +414,7 @@ public class XSDEditFacetAction extends UndoAction{
    		if (ret == Dialog.CANCEL) return;
 
    		if (currentValue!=null) std.getFacetContents().remove(currentValue);
-   		double intValue = Double.parseDouble(((InputDialog)dialog).getValue());
+   		int intValue = Integer.parseInt(((InputDialog)dialog).getValue());
    		if (intValue>0) {
    			XSDMaxInclusiveFacet f = (XSDSchemaBuildingTools.getXSDFactory()).createXSDMaxInclusiveFacet();
    			f.setLexicalValue(""+intValue);
@@ -448,7 +448,8 @@ public class XSDEditFacetAction extends UndoAction{
    		if (ret == Dialog.CANCEL) return;
 
    		if (currentValue!=null) std.getFacetContents().remove(currentValue);
-   		double intValue = Double.parseDouble(((InputDialog)dialog).getValue());
+//   		double intValue = Double.parseDouble(((InputDialog)dialog).getValue());
+   		int intValue = Integer.parseInt(((InputDialog)dialog).getValue());
    		if (intValue>0) {
    			XSDMaxExclusiveFacet f = (XSDSchemaBuildingTools.getXSDFactory()).createXSDMaxExclusiveFacet();
    			f.setLexicalValue(""+intValue);
@@ -483,7 +484,8 @@ public class XSDEditFacetAction extends UndoAction{
    		if (ret == Dialog.CANCEL) return;
 
    		if (currentValue!=null) std.getFacetContents().remove(currentValue);
-   		double intValue = Double.parseDouble(((InputDialog)dialog).getValue());
+   		int intValue = Integer.parseInt(((InputDialog)dialog).getValue());
+//   		double intValue = Double.parseDouble(((InputDialog)dialog).getValue());
    		if (intValue>0) {
    			XSDMinInclusiveFacet f = (XSDSchemaBuildingTools.getXSDFactory()).createXSDMinInclusiveFacet();
    			f.setLexicalValue(""+intValue);
@@ -518,7 +520,8 @@ public class XSDEditFacetAction extends UndoAction{
    		if (ret == Dialog.CANCEL) return;
 
    		if (currentValue!=null) std.getFacetContents().remove(currentValue);
-   		double intValue = Double.parseDouble(((InputDialog)dialog).getValue());
+   		int intValue = Integer.parseInt(((InputDialog)dialog).getValue());
+//   		double intValue = Double.parseDouble(((InputDialog)dialog).getValue());
    		if (intValue>0) {
    			XSDMinExclusiveFacet f = (XSDSchemaBuildingTools.getXSDFactory()).createXSDMinExclusiveFacet();
    			f.setLexicalValue(""+intValue);
